+++ .formatter.exs
@@ -1,0 +1,3 @@
+[
+  inputs: [".formatter.exs", "{src,test}/*.ex"]
+]